requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.1  
requests

